+++ Microsoft.Azure.Cosmos/src/Telemetry/VmMetadataApiHandler.cs
@@ -122,19 +122,20 @@ namespace Microsoft.Azure.Cosmos.Telemetry
         /// <returns>hashed Value</returns>
         internal static string ComputeHash(string rawData)
         {
+            if (string.IsNullOrEmpty(rawData))
+            {
+                throw new ArgumentNullException(nameof(rawData));
+            }
+
             // Create a SHA256    
             using (SHA256 sha256Hash = SHA256.Create())
             {
                 // ComputeHash - returns byte array  
                 byte[] bytes = sha256Hash.ComputeHash(Encoding.UTF8.GetBytes(rawData));
+                Array.Resize(ref bytes, 16);
 
                 // Convert byte array to a string   
-                StringBuilder builder = new StringBuilder();
-                for (int i = 0; i < bytes.Length; i++)
-                {
-                    builder.Append(bytes[i].ToString("x2"));
-                }
-                return builder.ToString();
+                return new Guid(bytes).ToString();
             }
         }
